Move job status functions from diracx-db to diracx-routers
